feat(interop): vizij-arora-tts — the say module shared as a crate
The TTS contract (say(text, voice) -> Status, the streaming viseme
out-parameter, the pasted ids) and the cloud provider move from the app
into a publishable crate on the host-module seam alone (arora-engine,
no features), so any arora host — the native app and the vizij-web
standalone alike — registers the same module. The app consumes the
crate (its Piper provider keeps implementing the same contract
app-side), and the say example drives the crate directly.
