+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
   "crates/interop/vizij-arora-hal",
   "crates/interop/vizij-arora-behavior",
   "crates/interop/vizij-arora-host",
+  "crates/interop/vizij-arora-tts",
   "crates/interop/vizij-piper",
   "crates/tools/vizij-bundle",
   "crates/tools/vizij-glb-migrate",
