+++ crates/vizij/Cargo.toml
@@ -65,13 +65,11 @@ vizij-arora-behavior = { path = "../interop/vizij-arora-behavior" }
 # dependency and no nightly. Its generated `Value` conversions marshal the calls.
 vizij-animation-module = { path = "../interop/vizij-animation-module" }
 
-# The TTS modules: `reqwest` reaches the Vizij TTS cloud provider over HTTP;
-# `rodio` (pure Rust, nothing to install) plays the audio whose playhead the
-# viseme cursor follows (both providers). `vizij-piper` is the local provider
-# (feature `tts-piper`); its build self-provisions libpiper and a patched
-# default voice, so no configuration is needed — note it links GPLv3 code
-# into the binary.
-reqwest = { version = "0.13", features = ["json"] }
+# The TTS contract + cloud provider, as a host module (shared with the
+# standalone). `rodio` plays the Piper provider's audio; `vizij-piper` is the
+# local provider (feature `tts-piper`) — its build self-provisions libpiper
+# and a patched default voice; note it links GPLv3 code into the binary.
+vizij-arora-tts = { version = "1", path = "../interop/vizij-arora-tts" }
 rodio = "0.20"
 vizij-piper = { path = "../interop/vizij-piper", optional = true }
